+++ CHANGELOG.md
@@ -6,6 +6,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fixed
+
+- Fixes `URI.encode` obsolete warning by using `Addressable` instead.
+
 ## [1.2.0] - 2020-07-24
 ### Added